# swagger.yaml
openapi: 3.0.3
info:
  title: Photo Studio Booking - API
  version: 1.1.0
  description: >
    API Gateway (5000) for booking & payments + integration endpoints.
    Payment webhook is served by Payment Service (5002).
    Message Broker (5050) и Notification Service (5004) описаны как отдельные servers.
servers:
  - url: http://localhost:5000
    description: API Gateway
  - url: http://localhost:5002
    description: Payment Service (webhooks)
  - url: http://localhost:5050
    description: Message Broker
  - url: http://localhost:5004
    description: Notification Service
paths:
...
  /api/payments:
    post:
      summary: Create payment
      tags: [Payments]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PaymentCreateRequest" }
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Payment" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/payments/{payment_id}:
    get:
      summary: Get payment by id
      tags: [Payments]
      parameters:
        - in: path
          name: payment_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Payment" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/payments/webhook/{gateway}:
    post:
      summary: Payment webhook (Payment Service 5002)
      description: >
        This endpoint is hosted on http://localhost:5002 (Payment Service),
        not via API Gateway. For YooKassa payload uses event/object structure.
      tags: [Payments]
      servers:
        - url: http://localhost:5002
      parameters:
        - in: path
          name: gateway
          required: true
          schema:
            type: string
            enum: [yookassa, sberpay, tinkoff]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/YooKassaWebhook" }
      responses:
        "200":
          description: Processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
        "400":
          description: Invalid gateway/payload
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Payment not found by external id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/payments/{payment_id}/refund:
    post:
      summary: Refund payment (only succeeded)
      description: >
        Before calling this endpoint, you must create a payment and process
        successful webhook so that payment status becomes `succeeded`.
      tags: [Payments]
      parameters:
        - in: path
          name: payment_id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefundRequest"
      responses:
        "200":
          description: Refund result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Refund" }
        "400":
          description: Invalid state or validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Выбранное время недоступно"
...
    PaymentCreateRequest:
      type: object
      required: [booking_id, amount, payment_type, payment_method]
      properties:
        booking_id: { type: string, format: uuid }
        amount: { type: number, example: 3000.0 }
        payment_type:
          type: string
          enum: [prepayment, full_payment]
          example: full_payment
        payment_method:
          type: string
          enum: [yookassa, sberpay, tinkoff]
          example: yookassa
        return_url:
          type: string
          nullable: true
          example: "https://example.com/return"
    Payment:
      type: object
      properties:
        payment_id: { type: string, format: uuid }
        booking_id: { type: string, format: uuid }
        amount: { type: string, example: "3000.0" }
        status:
          type: string
          enum: [pending, succeeded, failed, refunded]
          example: pending
        payment_method:
          type: string
          enum: [yookassa, sberpay, tinkoff]
        payment_url: { type: string, example: "https://yoomoney.ru/checkout/payments/v2/contract?orderId=..." }
        external_payment_id: { type: string, example: "yookassa-0294d913-3de5-4ad0-8616-4eb6b7b233a1" }
        created_at: { type: string, example: "Tue, 16 Dec 2025 23:39:28 GMT" }
        updated_at: { type: string, nullable: true, example: null }
    RefundRequest:
      type: object
      required: [payment_id]
      properties:
        payment_id:
          type: string
          format: uuid
          example: "af50a1ae-494a-465c-80af-446406a935ba"
        amount:
          type: number
          nullable: true
          example: 1500.0
        reason:
          type: string
          nullable: true
          example: "Partial refund"
    Refund:
      type: object
      properties:
        refund_id: { type: string, format: uuid }
        payment_id: { type: string, format: uuid }
        amount: { type: string, example: "1500.0" }
        status: { type: string, example: succeeded }
        timestamp: { type: string, format: date-time }
...
    YooKassaWebhook:
      type: object
      required: [event, object]
      properties:
        type: { type: string, example: notification }
        event: { type: string, example: payment.succeeded }
        object:
          type: object
          required: [id, status, amount]
          properties:
            id: { type: string, example: "yookassa-0294d913-3de5-4ad0-8616-4eb6b7b233a1" }
            status: { type: string, example: succeeded }
            amount:
              type: object
              properties:
                value: { type: string, example: "3000.0" }
                currency: { type: string, example: RUB }
            metadata:
              type: object
              additionalProperties: true
        created_at: { type: string, nullable: true, format: date-time }